language_app/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
supabase>=2.4.0
google-genai>=1.0.0
edge-tts>=7.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
openai>=1.30.0
